Move RandomSampler to sampler module
